[FEATURE] Slave violates NMT state tolerance
Previously, the NMT commands are handled in the user module,
where the addressed CN is not transitioned from ReadyToOperate to
Operational states within 5 POWERLINK cycles and signalized to
NMT_WRONG_STATE.

This commit implements processing of NMT commands in the kernel
module and only forwarding state change events to user module.
Now, addressed CN state transition are faster and processed
within 5 POWERLINK cycles.
