+++ stack/include/user/nmtmnu.h
@@ -43,6 +43,7 @@ SOFTWARE, EVEN IF ADVISED OF THE POSSIBILITY OF SUCH DAMAGE.
 // includes
 //------------------------------------------------------------------------------
 #include <common/oplkinc.h>
+#include <common/nmt.h>
 #include <user/nmtu.h>
 #include <oplk/event.h>
