+++ stack/src/user/nmt/nmtcnu.c
@@ -90,9 +90,6 @@ static tNmtCnuInstance   nmtCnuInstance_g;
 //------------------------------------------------------------------------------
 // local function prototypes
 //------------------------------------------------------------------------------
-static tNmtCommand getNmtCommand(tFrameInfo* pFrameInfo_p);
-static BOOL checkNodeIdList(BYTE* pbNmtCommandDate_p);
-static tOplkError commandCb(tFrameInfo* pFrameInfo_p);
 
 //============================================================================//
 //            P U B L I C   F U N C T I O N S                                 //
@@ -102,9 +99,9 @@ static tOplkError commandCb(tFrameInfo* pFrameInfo_p);
 /**
 \brief  Init nmtcnu module
 
-The function initializes an instance of the nmtcnu module
+The function initializes an instance of the nmtcnu module.
 
-\param  nodeId_p                Node ID of the local node
+\param  nodeId_p                Node ID of the local node.
 
 \return The function returns a tOplkError error code.
 
@@ -125,7 +122,7 @@ tOplkError nmtcnu_init(UINT nodeId_p)
     nmtCnuInstance_g.extNmtCmdBitMask = 1 << ((UINT8)nodeId_p & 7);
 
     // register callback-function for NMT-commands
-    ret = dllucal_regAsndService(kDllAsndNmtCommand, commandCb, kDllAsndFilterLocal);
+    ret = dllucal_regAsndService(kDllAsndNmtCommand, NULL, kDllAsndFilterLocal);
 
     return ret;
 }
@@ -250,268 +247,5 @@ tOplkError nmtcnu_registerCheckEventCb(tNmtuCheckEventCallback pfnNmtCheckEventC
 //============================================================================//
 /// \name Private Functions
 /// \{
-
-//------------------------------------------------------------------------------
-/**
-\brief  Callback function for NMT commands
-
-The function processes NMT commands.
-
-\param  pFrameInfo_p        Pointer to frame containing the NMT command
-
-\return The function returns a tOplkError error code.
-*/
-//------------------------------------------------------------------------------
-static tOplkError commandCb(tFrameInfo* pFrameInfo_p)
-{
-    tOplkError      ret = kErrorOk;
-    tNmtCommand     nmtCommand;
-    BOOL            fNodeIdInList;
-    tNmtEvent       nmtEvent = kNmtEventNoEvent;
-
-    if (pFrameInfo_p == NULL)
-        return kErrorNmtInvalidFramePointer;
-
-    nmtCommand = getNmtCommand(pFrameInfo_p);
-    switch (nmtCommand)
-    {
-        //------------------------------------------------------------------------
-        // plain NMT state commands
-        case kNmtCmdStartNode:
-            nmtEvent = kNmtEventStartNode;
-            break;
-
-        case kNmtCmdStopNode:
-            nmtEvent = kNmtEventStopNode;
-            break;
-
-        case kNmtCmdEnterPreOperational2:
-            nmtEvent = kNmtEventEnterPreOperational2;
-            break;
-
-        case kNmtCmdEnableReadyToOperate:
-            nmtEvent = kNmtEventEnableReadyToOperate;
-            break;
-
-        case kNmtCmdResetNode:
-            nmtEvent = kNmtEventResetNode;
-            break;
-
-        case kNmtCmdResetCommunication:
-            nmtEvent = kNmtEventResetCom;
-            break;
-
-        case kNmtCmdResetConfiguration:
-            nmtEvent = kNmtEventResetConfig;
-            break;
-
-        case kNmtCmdSwReset:
-            nmtEvent = kNmtEventSwReset;
-            break;
-
-        //------------------------------------------------------------------------
-        // extended NMT state commands
-        case kNmtCmdStartNodeEx:
-            // check if own nodeid is in the POWERLINK node list
-            fNodeIdInList = checkNodeIdList(&(pFrameInfo_p->frame.pBuffer->data.asnd.payload.nmtCommandService.aNmtCommandData[0]));
-            if (fNodeIdInList != FALSE)
-            {   // own nodeid in list
-                // send event to process command
-                nmtEvent = kNmtEventStartNode;
-            }
-            break;
-
-        case kNmtCmdStopNodeEx:
-            // check if own nodeid is in the POWERLINK node list
-            fNodeIdInList = checkNodeIdList(&pFrameInfo_p->frame.pBuffer->data.asnd.payload.nmtCommandService.aNmtCommandData[0]);
-            if (fNodeIdInList != FALSE)
-            {   // own nodeid in list
-                // send event to process command
-                nmtEvent = kNmtEventStopNode;
-            }
-            break;
-
-        case kNmtCmdEnterPreOperational2Ex:
-            // check if own nodeid is in the POWERLINK node list
-            fNodeIdInList = checkNodeIdList(&pFrameInfo_p->frame.pBuffer->data.asnd.payload.nmtCommandService.aNmtCommandData[0]);
-            if (fNodeIdInList != FALSE)
-            {   // own nodeid in list
-                // send event to process command
-                nmtEvent = kNmtEventEnterPreOperational2;
-            }
-            break;
-
-        case kNmtCmdEnableReadyToOperateEx:
-            // check if own nodeid is in the POWERLINK node list
-            fNodeIdInList = checkNodeIdList(&pFrameInfo_p->frame.pBuffer->data.asnd.payload.nmtCommandService.aNmtCommandData[0]);
-            if (fNodeIdInList != FALSE)
-            {   // own nodeid in list
-                // send event to process command
-                nmtEvent = kNmtEventEnableReadyToOperate;
-            }
-            break;
-
-        case kNmtCmdResetNodeEx:
-            // check if own nodeid is in the POWERLINK node list
-            fNodeIdInList = checkNodeIdList(&pFrameInfo_p->frame.pBuffer->data.asnd.payload.nmtCommandService.aNmtCommandData[0]);
-            if (fNodeIdInList != FALSE)
-            {   // own nodeid in list
-                // send event to process command
-                nmtEvent = kNmtEventResetNode;
-            }
-            break;
-
-        case kNmtCmdResetCommunicationEx:
-            // check if own nodeid is in the POWERLINK node list
-            fNodeIdInList = checkNodeIdList(&pFrameInfo_p->frame.pBuffer->data.asnd.payload.nmtCommandService.aNmtCommandData[0]);
-            if (fNodeIdInList != FALSE)
-            {   // own nodeid in list
-                // send event to process command
-                nmtEvent = kNmtEventResetCom;
-            }
-            break;
-
-        case kNmtCmdResetConfigurationEx:
-            // check if own nodeid is in the POWERLINK node list
-            fNodeIdInList = checkNodeIdList(&pFrameInfo_p->frame.pBuffer->data.asnd.payload.nmtCommandService.aNmtCommandData[0]);
-            if (fNodeIdInList != FALSE)
-            {   // own nodeid in list
-                // send event to process command
-                nmtEvent = kNmtEventResetConfig;
-            }
-            break;
-
-        case kNmtCmdSwResetEx:
-            // check if own nodeid is in the POWERLINK node list
-            fNodeIdInList = checkNodeIdList(&pFrameInfo_p->frame.pBuffer->data.asnd.payload.nmtCommandService.aNmtCommandData[0]);
-            if (fNodeIdInList != FALSE)
-            {   // own nodeid in list
-                // send event to process command
-                nmtEvent = kNmtEventSwReset;
-            }
-            break;
-
-        //------------------------------------------------------------------------
-        // NMT managing commands
-        // TODO: add functions to process managing command (optional)
-        case kNmtCmdNetHostNameSet:
-            break;
-
-        case kNmtCmdFlushArpEntry:
-            break;
-
-        //------------------------------------------------------------------------
-        // NMT info services
-        // TODO: forward event with infos to the application (optional)
-        case kNmtCmdPublishConfiguredCN:
-            break;
-
-        case kNmtCmdPublishActiveCN:
-            break;
-
-        case kNmtCmdPublishPreOperational1:
-            break;
-
-        case kNmtCmdPublishPreOperational2:
-            break;
-
-        case kNmtCmdPublishReadyToOperate:
-            break;
-
-        case kNmtCmdPublishOperational:
-            break;
-
-        case kNmtCmdPublishStopped:
-            break;
-
-        case kNmtCmdPublishEmergencyNew:
-            break;
-
-        case kNmtCmdPublishTime:
-            break;
-
-        //-----------------------------------------------------------------------
-        // error from MN
-        // -> requested command not supported by MN
-        case kNmtCmdInvalidService:
-            // TODO: errorevent to application
-            break;
-
-        //------------------------------------------------------------------------
-        // default
-        default:
-            return kErrorNmtUnknownCommand;
-            break;
-    } // end of switch (nmtCommand)
-
-    if (nmtEvent != kNmtEventNoEvent)
-    {
-        if (nmtCnuInstance_g.pfnCheckEventCb != NULL)
-        {
-            ret = nmtCnuInstance_g.pfnCheckEventCb(nmtEvent);
-            if (ret == kErrorReject)
-            {
-                return kErrorOk;
-            }
-            else if (ret != kErrorOk)
-            {
-                return ret;
-            }
-        }
-        ret = nmtu_postNmtEvent(nmtEvent);
-    }
-
-    return ret;
-}
-
-//------------------------------------------------------------------------------
-/**
-\brief  Get NMT command
-
-The function extracts the NMT command from the frame.
-
-\param  pFrameInfo_p        Pointer to frame containing the NMT command
-
-\return The function returns the extracted NMT command
-*/
-//------------------------------------------------------------------------------
-static tNmtCommand getNmtCommand(tFrameInfo* pFrameInfo_p)
-{
-    tNmtCommand          nmtCommand;
-    tNmtCommandService*  pNmtCommandService;
-
-    pNmtCommandService = &pFrameInfo_p->frame.pBuffer->data.asnd.payload.nmtCommandService;
-    nmtCommand = (tNmtCommand)ami_getUint8Le(&pNmtCommandService->nmtCommandId);
-
-    return nmtCommand;
-}
-
-//------------------------------------------------------------------------------
-/**
-\brief  Check node ID list
-
-The function checks if the own node ID is set in the node list.
-
-\param  pbNmtCommandDate_p        Pointer to the date of the NMT command.
-
-\return The function returns \b TRUE if the node is found in the node list or
-        \b FALSE if it is not found in the node list.
-*/
-//------------------------------------------------------------------------------
-static BOOL checkNodeIdList(UINT8* pbNmtCommandDate_p)
-{
-    BOOL            fNodeIdInList;
-    UINT            byteOffset = nmtCnuInstance_g.extNmtCmdByteOffset;
-    UINT8           bitMask = nmtCnuInstance_g.extNmtCmdBitMask;
-    UINT8           nodeListByte;
-
-    nodeListByte = ami_getUint8Le(&pbNmtCommandDate_p[byteOffset]);
-    if ((nodeListByte & bitMask) == 0)
-        fNodeIdInList = FALSE;
-    else
-        fNodeIdInList = TRUE;
-
-    return fNodeIdInList;
-}
 
 ///\}
